Ventas
validacion de clientes, proveedores, productos.
Arreglo de combo box en cascada.
implementacion de mask y de datepicker.
index de ventas, borrado y crear ventas.

+++ SantaMarta.Web/SantaMarta.WebAPI/Controllers/DetailController.cs
@@ -16,7 +16,7 @@ namespace SantaMarta.WebAPI.Controllers
         [HttpPost]
         public IHttpActionResult Post(Int64 id)
         {
-            int clients;
+            Int64 clients;
 
             DetailsB detailsB = new DetailsB();
